+++ op-challenger/fault/service.go
@@ -7,10 +7,12 @@ import (
 
 	"github.com/ethereum-optimism/optimism/op-bindings/bindings"
 	"github.com/ethereum-optimism/optimism/op-challenger/config"
+	"github.com/ethereum-optimism/optimism/op-challenger/fault/scheduler"
 	"github.com/ethereum-optimism/optimism/op-challenger/fault/types"
 	"github.com/ethereum-optimism/optimism/op-challenger/metrics"
 	"github.com/ethereum-optimism/optimism/op-challenger/version"
 	"github.com/ethereum-optimism/optimism/op-service/client"
+	"github.com/ethereum-optimism/optimism/op-service/clock"
 	oppprof "github.com/ethereum-optimism/optimism/op-service/pprof"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
 	"github.com/ethereum/go-ethereum/common"
@@ -18,21 +20,23 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 )
 
-// Service provides a clean interface for the challenger to interact
-// with the fault package.
-type Service interface {
-	// MonitorGame monitors the fault dispute game and attempts to progress it.
-	MonitorGame(context.Context) error
+// TODO(CLI-4342): Make this a cli option
+const maxConcurrency = 4
+
+type Loader interface {
+	FetchAbsolutePrestateHash(ctx context.Context) ([]byte, error)
 }
 
-type service struct {
+type Service struct {
 	logger  log.Logger
 	metrics metrics.Metricer
 	monitor *gameMonitor
+	sched   *scheduler.Scheduler
 }
 
 // NewService creates a new Service.
-func NewService(ctx context.Context, logger log.Logger, cfg *config.Config) (*service, error) {
+func NewService(ctx context.Context, logger log.Logger, cfg *config.Config) (*Service, error) {
+	cl := clock.SystemClock
 	m := metrics.NewMetrics()
 	txMgr, err := txmgr.NewSimpleTxManager("challenger", logger, &m.TxMetrics, cfg.TxMgrConfig)
 	if err != nil {
@@ -71,17 +75,25 @@ func NewService(ctx context.Context, logger log.Logger, cfg *config.Config) (*se
 	}
 	loader := NewGameLoader(factory)
 
-	monitor := newGameMonitor(logger, client.BlockNumber, cfg.GameAddress, loader, func(addr common.Address) (gamePlayer, error) {
-		return NewGamePlayer(ctx, logger, cfg, addr, txMgr, client)
-	})
+	disk := newDiskManager(cfg.Datadir)
+	sched := scheduler.NewScheduler(
+		logger,
+		disk,
+		maxConcurrency,
+		func(addr common.Address, dir string) (scheduler.GamePlayer, error) {
+			return NewGamePlayer(ctx, logger, cfg, dir, addr, txMgr, client)
+		})
+
+	monitor := newGameMonitor(logger, cl, loader, sched, cfg.GameWindow, client.BlockNumber, cfg.GameAllowlist)
 
 	m.RecordInfo(version.SimpleWithMeta)
 	m.RecordUp()
 
-	return &service{
+	return &Service{
 		logger:  logger,
 		metrics: m,
 		monitor: monitor,
+		sched:   sched,
 	}, nil
 }
 
@@ -103,6 +115,8 @@ func ValidateAbsolutePrestate(ctx context.Context, trace types.TraceProvider, lo
 }
 
 // MonitorGame monitors the fault dispute game and attempts to progress it.
-func (s *service) MonitorGame(ctx context.Context) error {
+func (s *Service) MonitorGame(ctx context.Context) error {
+	s.sched.Start(ctx)
+	defer s.sched.Close()
 	return s.monitor.MonitorGames(ctx)
 }