+++ op-e2e/e2eutils/challenger/helper.go
@@ -5,6 +5,7 @@ import (
 	"crypto/ecdsa"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
@@ -13,6 +14,7 @@ import (
 	op_challenger "github.com/ethereum-optimism/optimism/op-challenger"
 	"github.com/ethereum-optimism/optimism/op-challenger/config"
 	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils"
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/wait"
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum-optimism/optimism/op-node/testlog"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
@@ -23,9 +25,12 @@ import (
 )
 
 type Helper struct {
-	log    log.Logger
-	cancel func()
-	errors chan error
+	log     log.Logger
+	t       *testing.T
+	require *require.Assertions
+	dir     string
+	cancel  func()
+	errors  chan error
 }
 
 type Option func(config2 *config.Config)
@@ -38,7 +43,10 @@ func WithFactoryAddress(addr common.Address) Option {
 
 func WithGameAddress(addr common.Address) Option {
 	return func(c *config.Config) {
-		c.GameAddress = addr
+		if c.GameAllowlist == nil {
+			c.GameAllowlist = make([]common.Address, 0)
+		}
+		c.GameAllowlist = append(c.GameAllowlist, addr)
 	}
 }
 
@@ -71,7 +79,6 @@ func WithCannon(
 		require := require.New(t)
 		c.TraceType = config.TraceTypeCannon
 		c.CannonL2 = l2Endpoint
-		c.CannonDatadir = t.TempDir()
 		c.CannonBin = "../cannon/bin/cannon"
 		c.CannonServer = "../op-program/bin/op-program"
 		c.CannonAbsolutePreState = "../op-program/bin/prestate.json"
@@ -79,13 +86,13 @@ func WithCannon(
 
 		genesisBytes, err := json.Marshal(l2Genesis)
 		require.NoError(err, "marshall l2 genesis config")
-		genesisFile := filepath.Join(c.CannonDatadir, "l2-genesis.json")
+		genesisFile := filepath.Join(c.Datadir, "l2-genesis.json")
 		require.NoError(os.WriteFile(genesisFile, genesisBytes, 0644))
 		c.CannonL2GenesisPath = genesisFile
 
 		rollupBytes, err := json.Marshal(rollupCfg)
 		require.NoError(err, "marshall rollup config")
-		rollupFile := filepath.Join(c.CannonDatadir, "rollup.json")
+		rollupFile := filepath.Join(c.Datadir, "rollup.json")
 		require.NoError(os.WriteFile(rollupFile, rollupBytes, 0644))
 		c.CannonRollupConfigPath = rollupFile
 	}
@@ -103,9 +110,12 @@ func NewChallenger(t *testing.T, ctx context.Context, l1Endpoint string, name st
 		errCh <- op_challenger.Main(ctx, log, cfg)
 	}()
 	return &Helper{
-		log:    log,
-		cancel: cancel,
-		errors: errCh,
+		log:     log,
+		t:       t,
+		require: require.New(t),
+		dir:     cfg.Datadir,
+		cancel:  cancel,
+		errors:  errCh,
 	}
 }
 
@@ -118,6 +128,7 @@ func NewChallengerConfig(t *testing.T, l1Endpoint string, options ...Option) *co
 		AlphabetTrace:           "",
 		AgreeWithProposedOutput: true,
 		TxMgrConfig:             txmgrCfg,
+		Datadir:                 t.TempDir(),
 	}
 	for _, option := range options {
 		option(cfg)
@@ -151,4 +162,42 @@ func (h *Helper) Close() error {
 		}
 		return nil
 	}
+}
+
+type GameAddr interface {
+	Addr() common.Address
+}
+
+func (h *Helper) VerifyGameDataExists(games ...GameAddr) {
+	for _, game := range games {
+		addr := game.Addr()
+		h.require.DirExistsf(h.gameDataDir(addr), "should have data for game %v", addr)
+	}
+}
+
+func (h *Helper) WaitForGameDataDeletion(ctx context.Context, games ...GameAddr) {
+	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
+	defer cancel()
+	err := wait.For(ctx, time.Second, func() (bool, error) {
+		for _, game := range games {
+			addr := game.Addr()
+			dir := h.gameDataDir(addr)
+			_, err := os.Stat(dir)
+			if errors.Is(err, os.ErrNotExist) {
+				// This game has been successfully deleted
+				continue
+			}
+			if err != nil {
+				return false, fmt.Errorf("failed to check dir %v is deleted: %w", dir, err)
+			}
+			h.t.Errorf("Game data directory %v not yet deleted", dir)
+			return false, nil
+		}
+		return true, nil
+	})
+	h.require.NoErrorf(err, "should have deleted game data directories")
+}
+
+func (h *Helper) gameDataDir(addr common.Address) string {
+	return filepath.Join(h.dir, "game-"+addr.Hex())
 }